Refactor of js.
